ai_assistant/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
openai==1.3.7
python-dotenv==1.0.0
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
hypothesis==6.92.1
# Security and encryption
cryptography==41.0.7
# Knowledge base dependencies
numpy==1.24.3
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-multipart==0.0.6
aiofiles==23.2.0